Diagram now appears...but that's all it does so far
To display the diagram below, there is now a settings widget in the
sidebar that has a checkbox option to show the diagram below.  There is
still quite of bit of work to do (making the diagram look pretty,
getting to the same level of functionality as the old tool, etc) but I
wanted to commit the basic thing so others can work on it too.